requirements.txt
alembic==0.8.6
aniso8601==1.3.0
Authomatic==0.1.0.post1
boto==2.48.0
boto3==1.3.0
botocore==1.4.83
click==6.6
docutils==0.12
et-xmlfile==1.0.1
Flask==0.11.1
Flask-Cors==2.1.2
Flask-Migrate==1.8.0
Flask-OAuthlib==0.9.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
functools32==3.2.3.post2
futures==3.0.5
iso8601==0.1.11
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.5.1
lxml==3.6.0
M2Crypto==0.25.1
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.7.3
numpy==1.11.2
oauthlib==2.0.0
openpyxl==2.4.8
Pillow==3.4.1
psycopg2==2.6.1
py==1.4.34
pycrypto==2.6.1
PyJWT==1.4.2
pytest==3.2.5
pytest-env==0.6.2
python-dateutil==2.6.0
python-editor==1.0.1
pytz==2016.4
requests==2.9.1
requests-oauthlib==0.6.2
six==1.10.0
SQLAlchemy==1.0.13
typing==3.5.2.2
Werkzeug==0.11.10
git+ssh://git@github.com/Appen/lr_utilities.git@2017.11.3.2.23
